Text width, scss and some very not sure Gemfile changes.
focused on scss files. Tried to understand what cell width 1191 does, when max-width
for content should be 656px. changes some entries in Gemfile to use the original plugin,
that came with the theme.
Output in browser still has major problems. The articles, as shown on the root URL now only
use the 656px, for images as well. It should be only 656, when reading the body part.
featured_image and gallery_images should be allowed to be wider, both when looking at the
article on the root URL and when having clicked on the post and reading it on its unique
URL.