Deleted all lines of the local Gemfile and replaced them with the Gemfile on imac2017.
I think the plugins cause problems, as my Gemfile on Studio got extremely long, after
putting everythin that was in the Gemfile.lock into the Gemfile, when it told me, that
the Gemfile.lock was corrupted and had to be rebuilt using Bundle install. Questionable
construction technique from my side then trying to include everything that was in the
Gemfile.lock into the new Gemfile. The Gemfile was tiny, as it is now again.