First commit on draft of 05 questionnaire
Removed the _drafts/ directory completely from .gitignore and added it
to the tracked files.

The first version of 05_questionnaire.md was added to this commit.

All the output*.png files were moved to 'images/projects/output*.png',
where they should be accessible for the site building.

TODO: Create Header for 05_questionnaire.md