05Q added and controlled that all questions have ---
Used a vim regex expression and gc to replace <br> tags between
question and first line of answer with a --- (solid line, when rendered)

Controlled visually on the locally built jekyll site, that all
horizontal lines render as they should for all questions.

Next up is question by question checking, that everything is fine or if
something big needs to be done.

One commit per checked and possibly fixed question.